refactor: split friend functionality into its own service
